+++ lib/src/app_bar/app_bar.dart
@@ -12,7 +12,7 @@ import 'title.dart';
 /// An adapted version of [AppBar] that morphs while navigating.
 class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
   MorphingAppBar({
-    Key? key,
+    super.key,
     this.heroTag = 'MorphingAppBar',
     this.leading,
     this.automaticallyImplyLeading = true,
@@ -27,7 +27,6 @@ class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
     this.foregroundColor,
     this.iconTheme,
     this.actionsIconTheme,
-    this.textTheme,
     this.primary = true,
     this.centerTitle,
     this.excludeHeaderSemantics = false,
@@ -42,10 +41,10 @@ class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
   })  : assert(elevation == null || elevation >= 0.0),
         preferredSize = Size.fromHeight(
           toolbarHeight ?? kToolbarHeight + (bottom?.preferredSize.height ?? 0),
-        ),
-        super(key: key);
+        );
 
   /// Tag used for the internally created [Hero] widget.
+  // ignore: no-object-declaration
   final Object heroTag;
 
   /// See [AppBar.leading]
@@ -86,9 +85,6 @@ class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
 
   /// See [AppBar.actionsIconTheme]
   final IconThemeData? actionsIconTheme;
-
-  /// See [AppBar.textTheme]
-  final TextTheme? textTheme;
 
   /// See [AppBar.primary]
   final bool primary;
@@ -157,7 +153,6 @@ class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
         foregroundColor: foregroundColor,
         iconTheme: iconTheme,
         actionsIconTheme: actionsIconTheme,
-        textTheme: textTheme,
         primary: primary,
         centerTitle: centerTitle,
         excludeHeaderSemantics: excludeHeaderSemantics,
@@ -166,7 +161,6 @@ class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
         bottomOpacity: bottomOpacity,
         toolbarHeight: toolbarHeight,
         leadingWidth: leadingWidth,
-        backwardsCompatibility: false,
         toolbarTextStyle: toolbarTextStyle,
         titleTextStyle: titleTextStyle,
         systemOverlayStyle: actualSystemOverlayStyle,
@@ -175,7 +169,7 @@ class MorphingAppBar extends StatelessWidget implements PreferredSizeWidget {
   }
 
   Widget _flightShuttleBuilder(
-    BuildContext context,
+    BuildContext _,
     Animation<double> animation,
     HeroFlightDirection direction,
     BuildContext fromContext,
@@ -233,11 +227,8 @@ class _AnimatedAppBar extends AnimatedWidget {
       shape: state.shape,
       backgroundColor: state.backgroundColor,
       foregroundColor: state.foregroundColor,
-      // `brightness` is obsolete.
       iconTheme: state.iconTheme,
       actionsIconTheme: state.actionsIconTheme,
-      // We don't need to set `textTheme` as that's only used as fallback for
-      // `toolbarTextStyle` and `titleTextStyle`.
       // The value is the same for parent and child, so it doesn't matter which
       // one we use.
       primary: parent.appBar.primary,
@@ -248,7 +239,6 @@ class _AnimatedAppBar extends AnimatedWidget {
       bottomOpacity: state.bottomOpacity,
       toolbarHeight: state.toolbarHeight,
       leadingWidth: state.leadingWidth,
-      backwardsCompatibility: false,
       toolbarTextStyle: state.toolbarTextStyle,
       titleTextStyle: state.titleTextStyle,
       systemOverlayStyle: state.systemOverlayStyle,