+++ lib/src/app_bar/title.dart
@@ -13,33 +13,23 @@ class AnimatedTitle extends MultiChildRenderObjectWidget {
   AnimatedTitle(MorphingState state)
       : t = state.t,
         super(
-          children: [
-            _createChild(state.parent),
-            _createChild(state.child),
-          ],
+          children: [_createChild(state.parent), _createChild(state.child)],
         );
 
   final double t;
 
   static Widget _createChild(EndState state) {
     final title = state.appBar.title;
-    if (title == null) {
-      return SizedBox();
-    }
+    if (title == null) return const SizedBox();
 
-    var style = state.appBar.textTheme?.headline6 ??
-        state.appBarTheme.textTheme?.headline6 ??
-        state.theme.primaryTextTheme.headline6;
+    var style = state.titleTextStyle;
     if (style?.color != null) {
       style = style!.copyWith(color: style.color!.withOpacity(state.opacity));
     }
 
     return _AnimatedTitleParentDataWidget(
       hasLeading: state.leading != null,
-      child: DefaultTextStyle.merge(
-        style: style,
-        child: title,
-      ),
+      child: DefaultTextStyle.merge(style: style, child: title),
     );
   }
 
@@ -50,18 +40,17 @@ class AnimatedTitle extends MultiChildRenderObjectWidget {
   @override
   void updateRenderObject(
     BuildContext context,
-    covariant _AnimatedTitleLayout renderObject,
+    covariant RenderObject renderObject,
   ) =>
-      renderObject.t = t;
+      (renderObject as _AnimatedTitleLayout).t = t;
 }
 
 class _AnimatedTitleParentDataWidget
     extends ParentDataWidget<_AnimatedTitleParentData> {
   const _AnimatedTitleParentDataWidget({
-    Key? key,
     required this.hasLeading,
-    required Widget child,
-  }) : super(key: key, child: child);
+    required super.child,
+  });
 
   final bool hasLeading;
 
@@ -72,14 +61,11 @@ class _AnimatedTitleParentDataWidget
   void applyParentData(RenderObject renderObject) {
     assert(renderObject.parentData is _AnimatedTitleParentData);
     final parentData = renderObject.parentData! as _AnimatedTitleParentData;
-
     if (parentData.hasLeading == hasLeading) return;
 
     parentData.hasLeading = hasLeading;
     final targetParent = renderObject.parent;
-    if (targetParent is RenderObject) {
-      targetParent.markNeedsLayout();
-    }
+    if (targetParent is RenderObject) targetParent.markNeedsLayout();
   }
 }
 
@@ -89,7 +75,7 @@ class _AnimatedTitleParentData extends ContainerBoxParentData<RenderBox> {
 
 class _AnimatedTitleLayout
     extends AnimatedAppBarLayout<_AnimatedTitleParentData> {
-  _AnimatedTitleLayout({double t = 0}) : super(t: t);
+  _AnimatedTitleLayout({super.t});
 
   @override
   void setupParentData(RenderObject child) {
@@ -100,19 +86,16 @@ class _AnimatedTitleLayout
 
   @override
   double computeMinIntrinsicWidth(double height) =>
-      children.map<num>((c) => c.getMinIntrinsicWidth(height)).max.toDouble();
-
+      children.map((c) => c.getMinIntrinsicWidth(height)).max.toDouble();
   @override
   double computeMaxIntrinsicWidth(double height) =>
-      children.map<num>((c) => c.getMaxIntrinsicWidth(height)).max.toDouble();
-
+      children.map((c) => c.getMaxIntrinsicWidth(height)).max.toDouble();
   @override
   double computeMinIntrinsicHeight(double width) =>
-      children.map<num>((c) => c.getMinIntrinsicHeight(width)).max.toDouble();
-
+      children.map((c) => c.getMinIntrinsicHeight(width)).max.toDouble();
   @override
   double computeMaxIntrinsicHeight(double width) =>
-      children.map<num>((c) => c.getMaxIntrinsicHeight(width)).max.toDouble();
+      children.map((c) => c.getMaxIntrinsicHeight(width)).max.toDouble();
 
   @override
   bool get alwaysNeedsCompositing => true;
@@ -148,12 +131,12 @@ class _AnimatedTitleLayout
     context
       ..pushOpacity(
         parent.data.offset + offset,
-        math.max<double>(0, 1 - 2 * t).opacityToAlpha,
+        math.max<double>(0, 1 - t * 2).opacityToAlpha,
         (context, offset) => context.paintChild(parent, offset),
       )
       ..pushOpacity(
         child.data.offset + offset,
-        math.max<double>(0, 2 * t - 1).opacityToAlpha,
+        math.max<double>(0, t * 2 - 1).opacityToAlpha,
         (context, offset) => context.paintChild(child, offset),
       );
   }